+++ V2rayNG/app/src/main/kotlin/com/v2ray/ang/service/ServiceControl.kt
@@ -1,0 +1,15 @@
+package com.v2ray.ang.service
+
+import android.app.Service
+
+interface ServiceControl {
+    fun getService(): Service
+
+    fun startService(parameters: String)
+
+    fun stopService()
+
+    fun vpnProtect(socket: Int): Boolean
+
+    fun vpnSendFd()
+}
